refactor(combat): ActionCostValidator の mpCost を int に一本化
ActionCostValidator.CanAfford / DeductCost の mpCost パラメータを float から int に変更し、
内部の (int)mpCost キャストを削除。合わせて AttackInfo.mpCost / AttackMotionData.mpCost も
int に揃え、ActionExecutorController.CheckAndDeductCost のシグネチャも更新。

- Assets/MyAsset/Core/Combat/ActionCostValidator.cs: int 化
- Assets/MyAsset/Core/ScriptableObjects/AttackInfo.cs: mpCost を int に
- Assets/MyAsset/Core/Common/Structs.cs (AttackMotionData): mpCost を int に
- Assets/MyAsset/Runtime/Combat/ActionExecutorController.cs: CheckAndDeductCost 引数型変更
- 既存テスト (ActionExecutorControllerTests / CombatDataHelperTests /
  Integration_ActionExecutorRuntimeTests) の f サフィックスを除去
- 境界値テスト 3 本追加 (== 判定 / ジャスト消費 / 超過時 0 クランプ)
- FUTURE_TASKS.md 該当項目を int 一本化でチェック済み化

備考: SkillConditionLogic / ItemUsageLogic / CompanionMpManager は独立ドメインのため
今回の変更対象外。将来的に全体統合する場合は別タスクで扱う。

+++ Assets/MyAsset/Core/ScriptableObjects/AttackInfo.cs
@@ -51,7 +51,7 @@ namespace Game.Core
 
         [TitleGroup("コスト")]
         [MinValue(0)]
-        public float mpCost;
+        public int mpCost;
 
         [TitleGroup("コスト")]
         [MinValue(0)]
